Initial fix change and move to a model folder
