+++ src/main/java/com/biotrio/nocristina/models/Movie.java
@@ -1,4 +1,4 @@
-package com.biotrio.nocristina.movies;
+package com.biotrio.nocristina.models;
 
 public class Movie {
